Test Entity with null exception

+++ TravelAgency-persistence/src/main/java/cz/muni/fi/travelAgency/dao/ReservationDaoImp.java
@@ -10,8 +10,7 @@ import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.persistence.TypedQuery;
 import java.util.Collection;
-import java.util.Date;
-
+import java.time.LocalDate;
 /**
  * Implementation of {@link ReservationDao}
  *
@@ -59,7 +58,7 @@ public class ReservationDaoImp implements ReservationDao {
     }
 
     @Override
-    public Collection<Reservation> findReservationBetween(Date startDate, Date endDate) {
+    public Collection<Reservation> findReservationBetween(LocalDate startDate, LocalDate endDate) {
         TypedQuery<Reservation> query = em.createQuery("SELECT e FROM Reservation e WHERE e.reserveDate BETWEEN :startDate AND :endDate",
                 Reservation.class);
         query.setParameter("startDate", startDate);
@@ -69,11 +68,22 @@ public class ReservationDaoImp implements ReservationDao {
 
     @Override
     public void update(Reservation reservation) {
-        em.merge(reservation);
+        if (reservation == null) throw new IllegalArgumentException("Tried to update NULL!");
+        if (em.find(Reservation.class, reservation.getId()) != null){
+            em.merge(reservation);
+        } else {
+            throw new IllegalArgumentException("Tried to update Reservation that was not saved before.");
+        }
     }
 
     @Override
     public void remove(Reservation reservation) {
-        em.remove(em.merge(reservation));
+        if (reservation == null) throw new IllegalArgumentException("Tried to delete NULL!");
+        if (em.find(Reservation.class, reservation.getId()) != null){
+            em.remove(em.merge(reservation));
+        } else {
+            throw new IllegalArgumentException("Tried to Remove Reservation that was not saved before.");
+        }
+
     }
 }
